+++ src/ListRenderer/OrderedListRenderer.ts
@@ -1,6 +1,6 @@
 import type { OrderedListItemMeta } from "../types/IItemMeta";
 import { NestedListConfig } from "../types/LiistParams";
-import * as Dom from '../utils/dom';
+import * as Dom from '../utils/DDom';
 import { ListRenderer } from './ListRenderer';
 import type { ListCssClasses } from './ListRenderer';
