+++ src/index.ts
@@ -4,7 +4,7 @@ import type {
   BlockToolConstructorOptions,
   TunesMenuConfig,
 } from '@editorjs/editorjs/types/tools';
-import Caret from './utils/caret';
+import Caret from './utils/Caret';
 import { IconListBulleted, IconListNumbered, IconChecklist } from '@codexteam/icons';
 import { NestedListConfig, ListData, ListDataStyle, ListItem } from './types/LiistParams';
 import ListTabulator from './ListTabulator';
