--- src/types/IIcons.d.ts
@@ -1,7 +1,0 @@
-// temporary fix for the missing types
-declare module '@codexteam/icons' {
-  export const IconListBulleted: string;
-  export const IconListNumbered: string;
-  export const IconChecklist: string
-  export const IconCheck: string;
-}
