+++ src/utils/caret.ts
@@ -1,4 +1,4 @@
-import * as dom from './dom';
+import * as dom from './DDom';
 import { isHtmlElement } from './type-guards';
 
 /**
